Fixed sending cookies when user is unauthorized
