+++ src/server/passport/passportStatics.js
@@ -1,0 +1,3 @@
+module.exports = {
+	LOGIN_STRATEGY: 'LOGIN_STRATEGY'
+}
